+++ platforms/cuda/include/CudaNonbondedSlicingKernels.h
@@ -13,7 +13,7 @@
 
 #include "CommonNonbondedSlicingKernels.h"
 #include "NonbondedSlicingKernels.h"
-#include "CudaCuFFT3D.h"
+#include "internal/CudaCuFFT3D.h"
 #include "openmm/cuda/CudaContext.h"
 #include "openmm/Platform.h"
 #include "openmm/System.h"
